+++ packages/ice/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 3.4.0
+
+### Minor Changes
+
+- a68ff48d: feat: support @ice/pack-binding
+
+### Patch Changes
+
+- 7d193fe9: fix: css module hash and HMR
+- Updated dependencies [7d193fe9]
+- Updated dependencies [7d193fe9]
+- Updated dependencies [a68ff48d]
+  - @ice/rspack-config@1.1.0
+  - @ice/webpack-config@1.1.7
+  - @ice/shared-config@1.2.0
+  - @ice/bundles@0.2.0
+
 ## 3.3.9
 
 ### Patch Changes